Fixes pwd failing when in a directory that does not exists
